+++ src/frameworks/index.ts
@@ -34,9 +34,10 @@ import {
   VALID_ENGINES,
   WebFrameworks,
 } from "./constants";
-import { FirebaseDefaults, Framework } from "./interfaces";
+import { BuildResult, FirebaseDefaults, Framework } from "./interfaces";
 import { logWarning } from "../utils";
 import { ensureTargeted } from "../functions/ensureTargeted";
+import { isDeepStrictEqual } from "util";
 
 export { WebFrameworks };
 
@@ -65,6 +66,21 @@ export async function discover(dir: string, warn = true) {
   }
   if (warn) console.warn("Could not determine the web framework in use.");
   return;
+}
+
+const BUILD_MEMO = new Map<string[], Promise<BuildResult | void>>();
+
+// Memoize the build based on both the dir and the environment variables
+function memoizeBuild(dir: string, build: (dir: string) => Promise<BuildResult | void>, deps: any[]) {
+  const key = [dir, ...deps];
+  for (const existingKey of BUILD_MEMO.keys()) {
+    if (isDeepStrictEqual(existingKey, key)) {
+      return BUILD_MEMO.get(existingKey);
+    }
+  }
+  const value = build(dir);
+  BUILD_MEMO.set(key, value);
+  return value;
 }
 
 /**
@@ -129,6 +145,7 @@ export async function prepareFrameworks(
       throw new Error(`hosting.public and hosting.source cannot both be set in firebase.json`);
     }
     const ssrRegion = frameworksBackend?.region ?? DEFAULT_REGION;
+    const omitCloudFunction = frameworksBackend?.omit ?? false;
     if (!allowedRegionsValues.includes(ssrRegion)) {
       const validRegions = conjoinOptions(allowedRegionsValues);
       throw new FirebaseError(
@@ -252,7 +269,7 @@ export async function prepareFrameworks(
         headers = [],
         trailingSlash,
         i18n,
-      } = (await build(getProjectPath())) || {};
+      } = (await memoizeBuild(getProjectPath(), build, [firebaseDefaults])) || {};
 
       config.rewrites.push(...rewrites);
       config.redirects.push(...redirects);
@@ -269,7 +286,7 @@ export async function prepareFrameworks(
       });
 
       config.public = relative(projectRoot, hostingDist);
-      if (wantsBackend) codegenFunctionsDirectory = codegenProdModeFunctionsDirectory;
+      if (wantsBackend && !omitCloudFunction) codegenFunctionsDirectory = codegenProdModeFunctionsDirectory;
     }
     config.webFramework = `${framework}${codegenFunctionsDirectory ? "_ssr" : ""}`;
     if (codegenFunctionsDirectory) {
@@ -467,6 +484,9 @@ ${firebaseDefaults ? `__FIREBASE_DEFAULTS__=${JSON.stringify(firebaseDefaults)}\
         );
       }
     } else {
+      if (await pathExists(functionsDist)) {
+        await rm(functionsDist, { recursive: true });
+      }
       // No function, treat as an SPA
       // TODO(jamesdaniels) be smarter about this, leave it to the framework?
       config.rewrites.push({
@@ -491,6 +511,9 @@ ${firebaseDefaults ? `__FIREBASE_DEFAULTS__=${JSON.stringify(firebaseDefaults)}\
       });
     }
   }
+
+  // Clean up memos/caches
+  BUILD_MEMO.clear();
 }
 
 function codegenDevModeFunctionsDirectory() {